Add dry-run option

+++ src/lib/client.rb
@@ -709,6 +709,7 @@ class Client
       :nodes => [],
       :get_version => false,
       :get_user_info => false,
+      :dry_run => false,
       :multi_server => false,
       :servers => load_configfile(),
       :chosen_server => nil,
@@ -743,6 +744,9 @@ class Client
       opt.on("-S","--server STRING", "Specify the Kadeploy server to use") { |s|
         options[:chosen_server] = s
       }
+      opt.on("","--[no-]dry-run", "Perform a dry run") { |v|
+        options[:dry_run] = v
+      }
       opt.separator ""
       yield(opt,options)
     end
@@ -760,6 +764,12 @@ class Client
     options[:server_port] = options[:servers][options[:chosen_server]][1]
 
     return options
+  end
+
+  def init_params(options)
+    ret = { :user => USER }
+    ret[:dry_run] = options[:dry_run] if options[:dry_run]
+    ret
   end
 
   def self.launch()
@@ -804,7 +814,8 @@ class Client
     $clients.each do |client|
       $threads << Thread.new do
         Thread.current[:client] = client
-        client.run(options)
+        ret = client.run(options)
+        client.result(options,ret) unless options[:dry_run]
       end
     end
     $threads.each { |thread| thread.join }
@@ -833,6 +844,9 @@ class Client
 
   def run(options)
     raise
+  end
+
+  def result(options,ret)
   end
 
   def self.operation()
@@ -1063,8 +1077,15 @@ class ClientWorkflow < Client
     true
   end
 
+  def init_params(options)
+    super(options).merge({
+      :nodes => nodes(),
+    })
+  end
+
   def run_workflow(options,params)
     ret = post(api_path(),params.to_json)
+    return if options[:dry_run]
     @wid = ret['wid']
     @resources = ret['resources']
     File.open(options[:wid_file],'w'){|f| f.write @wid} if options[:wid_file]
@@ -1108,6 +1129,12 @@ p @resources['resource']
 
     debug "#{self.class.operation()}#{" ##{@wid}" if @wid} done\n\n"
 
+    delete(api_path()) if @wid
+
+    res
+  end
+
+  def result(options,res)
     unless res['error']
       # Success
       if res['nodes']['ok'] and !res['nodes']['ok'].empty?
@@ -1121,8 +1148,6 @@ p @resources['resource']
         debug res['nodes']['ko'].join("\n")
       end
     end
-
-    delete(api_path()) if @wid
   end
 end
